// types/appwrite.d.ts
import { type Models } from 'appwrite';

// This file is auto-generated by the Appwrite CLI. 
// You can regenerate it by running `appwrite types types/`.

export type Credentials = Models.Row & {
    userId: string;
    publicKey: string;
    counter: number;
    createdAt: string | null;
    updatedAt: string | null;
}

export type Challenges = Models.Row & {
    userId: string;
    challenge: string;
    createdAt: string | null;
}